Refactor : util 함수 분리
- util/commonFunction.tsx 파일에 비즈니스 로직 분리
