+++ src/components/CountryInfo.tsx
