+++ src/components/InputCard.tsx
