+++ src/components/chart/ChartCard.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import BasicCard from '../BasicCard';
 import Chart from './Chart';
 import CountryInfo from '../CountryInfo';
-import europe from '../../../assets/images/europe.png';
+import europe from '../../assets/images/europe.png';
 
 interface ParamProps {
   basePrice: number;
