+++ src/components/price_info/DataCard.tsx
