+++ src/interface/interface.tsx
@@ -1,0 +1,39 @@
+interface IData {
+  basePrice: number;
+  cashBuyingPrice: number;
+  cashSellingPrice: number;
+  changetime: string;
+  changePrice: number;
+  changeRate: number;
+  codetime: string;
+  countrytime: string;
+  createdAttime: string;
+  currencyCodetime: string;
+  currencyNametime: string;
+  currencyUnit: number;
+  datetime: string;
+  exchangeCommission: number;
+  fcSellingPricetime: null;
+  high52wDatetime: string;
+  high52wPrice: number;
+  highPrice: number;
+  id: number;
+  low52wDatetime: string;
+  low52wPrice: number;
+  lowPrice: number;
+  modifiedAttime: string;
+  nametime: string;
+  openingPrice: number;
+  providertime: string;
+  recurrenceCount: number;
+  signedChangePrice: number;
+  signedChangeRate: number;
+  tcBuyingPrice: null;
+  time: string;
+  timestamp: number;
+  ttBuyingPrice: number;
+  ttSellingPrice: number;
+  usDollarRate: number;
+}
+
+export default IData;
